Rename MaxPeople to MaxPart
